sys/vfs: reserve file descriptors for stdin, stdout, and stderr
Previously, this was done in uart_stdio. However, this does not work when
rtt_stdio is used instead. By moving the code into the VFS module, it can be
used regardless of the mechanism to write characters to the screen.

Adding a vfs_init() function also allows us to initialize the_vfs_open_files
array, which previously was not being done.
